Add a special AJAX endpoint for Frontend calls
If you need the FrontendUser object in your Ajax call
you have to set a lot of constants on your own. To prevent
that i have added a new class SimpleAjaxFrontend to give
the developer an easier way to do that.

There is also a new hook for Frontend-only AJAX calls. But
you can use both hooks if you want.

Signed-off-by: Leo Unglaub <leo@leo-unglaub.net>